==> p2/Part1.ipynb <==
import string
from tqdm import tqdm as timer
import re
import copy
import pandas as pd
import pickle as pk

# Filter Data

## Step 1 : Iterate through the input files, discarding lines with fewer than 10 movies. Generate a actor to movies mapping, and a reverse movies to actor mapping

The lines are cleaned by:
- Removing unprintable characters
- Removing stuff in brackets () or {}

A = {}
M = {}

lc = {"data/actress_movies.txt" : 1182813, "data/actor_movies.txt" : 2167653}

printable = set(string.printable)

def clean_string(s):
    # Remove (*)
    s = re.sub(r'\([^0-9]+\)', '', s)

    # Remove {*}
    s = re.sub(r'\{.*\}', '', s)
    
    # Remove multiple spaces
    s = re.sub(r'\s+', ' ', s)

    # Strip spaces etc
    s = s.lstrip()
    s = s.rstrip()

    return s

for fname in ["data/actor_movies.txt", "data/actress_movies.txt"]:
    with open(fname, "r") as f:
        for line in timer(f, total = lc[fname], desc=fname.split("/")[1]):
            if line.count('\t\t') <= 10:
                pass
            else:
                line = filter(lambda x : x in printable, line.decode('latin1')).encode('ascii')
                splits = line.split("\t\t")
                actor_name = splits[0]
                movies = set(map(clean_string, splits[1:]))

                if actor_name in A:
                    A[actor_name] = A[actor_name].union(movies)
                else:
                    A[actor_name] = movies

                for movie in movies:
                    if movie not in M:
                        M[movie] = [actor_name]
                    else:
                        M[movie].append(actor_name)

Define a relaxation to be one iteration of removing:
- Movies with less than 10 actors (update actors accordingly)
- Actors with less than 10 movies (update movies accordingly)

Clearly this is a cyclic process. We notice that the actors and their movies stabilize in about 33 iterations!

# Start throwing movies away
def relax(A, M):
    # Don't throw Rogue nation out
    extras = M["Mission: Impossible - Rogue Nation (2015)"]
    
    pop_movies = set(filter(lambda x : len(M[x]) > 10, M.keys()))
    A = {actor : set(movies).intersection(pop_movies) for (actor, movies) in A.iteritems()}
    A = {actor : movies for (actor, movies) in A.iteritems() if len(movies) > 10}
    
    good_actors = set(A.keys())
    M = {movie : set(actors).intersection(good_actors) for (movie, actors) in M.iteritems()}
    M = {movie : actors for (movie, actors) in M.iteritems() if len(actors) > 10}
    M["Mission: Impossible - Rogue Nation (2015)"] = filter(lambda x : x in A, extras)
    print len(A.keys()), len(M.keys())
    
    print len(A.keys()), len(M.keys())
    return A, M

for i in range(33):
    A, M = relax(A, M)

# Number of actors
len(A.keys())

# Number of movies
len(M.keys())

# Save A and M for later
with open("pyfiles/A.pkl", "w") as afile, open("pyfiles/M.pkl", "w") as mfile:
    pk.dump(A, afile)
    pk.dump(M, mfile)

# Write to file for next phase

with open('movie_map.txt', 'w') as f:
    for (actor, movies) in A.iteritems():
        f.write("%s \t\t %s \n" % (actor, '\t\t'.join(movies)))

# Clean the ratings

R = {}

printable = set(string.printable)

def clean_string(s):
    # Remove (*)
    s = re.sub(r'\([^0-9]+\)', '', s)

    # Remove {*}
    s = re.sub(r'\{.*\}', '', s)
    
    # Remove multiple spaces
    s = re.sub(r'\s+', ' ', s)

    # Strip spaces etc
    s = s.lstrip()
    s = s.rstrip()

    return s

with open("data/movie_rating.txt", "r") as fin, open("ratings.txt", "w") as fout:
    for line in fin:
        line  = filter(lambda x : x in printable, line.decode('latin1')).encode('ascii')
        
        movie, rating = line.split('\t\t')
        
        movie = clean_string(movie)
        rating = float(rating)
        
        if movie not in R:
            R[movie] = rating
    
    del R['']
    
    for (movie, rating) in R.iteritems():
        fout.write("%s\t%f\n" % (movie, rating))
