==> p2/Part3.ipynb <==
library(data.table)
library(igraph)

Run pagerank algorithm on the actor/actress network, look into those who are
among top 10, do you know their names? List the top 10 famous movie celebrities
in your opinion, what are their pagerank scores? Do you see any significant
pairings among actors? Any major surprises, in the sense that well-known actors
do not show up in the high pagerank list?

# Read Graph from DataFrame

gframe = read.table(file="elist.txt", header = FALSE, sep = "\t", quote="", dec = ".")
colnames(gframe) = c("node_1", "node_2", "weight")
g = graph.data.frame(gframe, directed=TRUE)

# Run a page rank

res = unlist( page.rank(g, weights = E(g)$weight) )

table = data.table(Name = V(g)$name, PageRank = res[1 : vcount(g)])

table = table [ order(table$PageRank, decreasing = TRUE) ]

head(table, 10)

write.table(table, "page_rank.txt", sep="\t", quote=FALSE, row.names = FALSE)
