==> p2/Part6.ipynb <==
Add the following nodes into the network, For each of them, return the top 5
nearest neighbors. Which communities does each of them belong to?
- Batman v Superman: Dawn of Justice (2016)
- Mission: Impossible - Rogue Nation (2015)
- Minions (2015)

library(data.table)
library(igraph)

# Load the graph from Part 5

g = readRDS("rfiles/g.rds")

get_nearest = function(v) {
    ns = neighbors(g, v)
    ws = sapply(ns, function(x){ edge_attr(g, "weight", get.edge.ids(g, c(v, x))) })
    return( ns[ order(ws, decreasing = TRUE) ] )
}

# Batman v Superman: Dawn of Justice (2016)

ns = get_nearest( which(V(g)$name == "Batman v Superman: Dawn of Justice (2016)") )
data.table(name = V(g)$name[ns], genre = V(g)$genre, community = V(g)$comm.id[ns])[1 : 10, ]

# Mission: Impossible - Rogue Nation (2015)
We are using "Mission: Impossible III (2006)  Thriller: instead, since it's in the graph. (report pe mat dalna)

ns = get_nearest( which(V(g)$name == "Mission: Impossible - Rogue Nation (2015)") )
data.table(name = V(g)$name[ns], genre = V(g)$genre, community = V(g)$comm.id[ns])[1 : 10, ]

# Minions (2015)

ns = get_nearest( which(V(g)$name == "Minions (2015)") )
data.table(name = V(g)$name[ns], genre = V(g)$genre, community = V(g)$comm.id[ns])[1 : 10, ]
