==> p2/Part7.ipynb <==
library(data.table)
library(igraph)

 Download the ratings list, derive a function to predict the ratings of the above 3
movies using the movie network. (hint: try to use the ratings of neighbor movies
and movies in the same community.)

# Read in graph from Part 5

g = readRDS("rfiles/g.rds")

ratings = fread("ratings.txt")
setkey(ratings, V1)

V(g)$rating = ratings[V(g)$name, ]$V2

# Clearly a whole bunch of ratings are missing!
sum( is.na( V(g)$rating ) )

movies = c(
    "Batman v Superman: Dawn of Justice (2016)",
    "Mission: Impossible - Rogue Nation (2015)",
    "Minions (2015)"
)

vids = sapply(movies, function(x){ which( V(g)$name == x) })

get_neighbors = function(v, k = NULL) {
    ns = neighbors(g, v)
    
    if(is.null( k )){
        return( ns )
    }
    else{
        ws = sapply(ns, function(x){ edge_attr(g, "weight", get.edge.ids(g, c(v, x))) })
        ns[ order(ws, decreasing = TRUE) ][1 : min(length(ns), k)]
    }
}

get_nearest_neighbor_average = function(x){
    nns = get_neighbors(x, 20)
    mean( na.omit( V(g)$rating[nns] ) )
}

get_all_neighbor_average = function(x){
    ns = get_neighbors(x)
    mean( na.omit( V(g)$rating[ns] ) )
}

actual_ratings = c(6.7, 6.8, 6.4)

community_average_ratings = sapply(vids, function(x){
    mean( na.omit( V(g)$ratings[ which( V(g)$comm.id[x] == V(g)$comm.id ) ] ) )
})

nearest_neighbor_average_ratings = sapply(vids, function(x){
    mean( na.omit( V(g)$rating[get_neighbors(x, 20)] ) )
})

all_neighbor_average_ratings = sapply(vids, function(x){
    mean( na.omit( V(g)$rating[get_neighbors(x)] ) )
})

combined_ratings = community_average_ratings + nearest_neighbor_average_ratings + all_neighbor_average_ratings
combined_ratings = sapply(combined_ratings, function(x){ x / 3})

data.table(
    Movie = movies,
    Actual = actual_ratings,
    Community = community_average_ratings,
    `Nearest Neighbors` = nearest_neighbor_average_ratings,
    `All Neighbors` = all_neighbor_average_ratings,
    Combined = combined_ratings
)
